chore(server): restructred commandParser service to be a directory for organization
